+++ ThirdPartyAdapters/bidmachine/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## BidMachine Android Mediation Adapter Changelog
 
+#### Next Version
+- Adds support for waterfall interstitial ads.
+
 #### Version 3.4.0.0 (In progress)
 - Added `MediationAdCallback.onAdOpened()` call when BidMachine's `onAdImpression()` callback is received for full-screen ads. This ensures that publishers recieve the correct callbacks when full-screen ad is displayed.